requirements.txt
click==8.1.3
Flask==2.2.3
Flask-Cors==3.0.10
Flask-JWT-Extended==4.4.4
Flask-SQLAlchemy==3.0.3
Flask-UUID==0.2
greenlet==2.0.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
passlib==1.7.4
PyJWT==2.6.0
python-dotenv==0.21.1
six==1.16.0
SQLAlchemy==2.0.4
typing_extensions==4.5.0
Werkzeug==2.2.3
